comment api

+++ src/main/java/com/example/communication/controller/PublishController.java
@@ -31,7 +31,7 @@ public class PublishController {
             @RequestParam(value = "title", required = false) String title,
             @RequestParam(value = "description", required = false) String description,
             @RequestParam(value = "tag", required = false) String tag,
-            @RequestParam(value = "id", required = false) Integer id,
+            @RequestParam(value = "id", required = false) Long id,
             HttpServletRequest request,
             Model model) {
 
@@ -81,7 +81,7 @@ public class PublishController {
     }
 
     @GetMapping("/publish/{id}")
-    public String publishUpdate(@PathVariable(value = "id") Integer id,
+    public String publishUpdate(@PathVariable(value = "id") Long id,
                                 Model model) {
         QuestionDto questionDto = questionDtoService.selectById(id);
         model.addAttribute("title", questionDto.getTitle());
